feat: Implement chat full screen mode - MEED-9674 - Meeds-io/MIPs#225 (#343)
Implement chat full screen mode

feat: Enhance chat fullpage mode ui/ux - MEED-9674 - Meeds-io/MIPs#225 (#347)

Enhance chat fullpage mode ui/ux
