+++ webapp/src/main/webapp/vue-apps/matrix/components/MatrixChatRooms.vue
@@ -5,6 +5,7 @@
     <matrix-chat-room
       v-for="room in rooms"
       :key="room.id"
+      :selectedRoom="selectedRoom"
       :room="room" />
   </div>
   <div v-else-if="!loading" class="d-flex full-height align-center justify-center full-width">
@@ -28,6 +29,10 @@
       loading: {
         type: Boolean,
         default: false
+      },
+      selectedRoom: {
+        type: Object,
+        default: null
       }
     },
     created() {
