--- webapp/src/main/webapp/vue-apps/matrix/components/component.vue
@@ -1,31 +1,0 @@
-<template>
-  <span>
-      {{ $t('matrix.description') }}
-  </span>
-</template>
-<script>
-  export default {
-    props: {
-      
-    },
-    data: () => ({
-    }),
-    created() {
-      this.initMatrixRoom();
-    },
-    methods: {
-      initMatrixRoom() {
-        return fetch(`${eXo.env.portal.context}/${eXo.env.portal.rest}/matrix?spaceId=${this.identityId}`, {
-          method: 'GET',
-          credentials: 'include',
-        }).then(resp => {
-          if (!resp || !resp.ok) {
-            console.warn('Could not get the URL of the team !');
-          } else {
-            return resp.text();
-          }
-        });
-      }
-    }
-  };
-</script>
